Rearrange config tests to separate modules
Rearranges the tests to better separate these tests to the individual
module it tests. Also edits a helper function in the command_line config
module to better allow for further testing with different example
command line arguments.
