+++ splinterd/src/config/toml.rs
@@ -96,3 +96,102 @@ pub fn from_file(mut f: File) -> Result<PartialConfig, ConfigError> {
 
     toml::from_str::<PartialConfig>(&toml).map_err(ConfigError::from)
 }
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    use std::fs;
+
+    /// Path to existing example config toml files from the top-level Splinterd directory.
+    static TEST_TOML: &str = "sample_configs/splinterd.toml.example";
+
+    /// Values present in the existing example config TEST_TOML file.
+    static EXAMPLE_STORAGE: &str = "yaml";
+    static EXAMPLE_TRANSPORT: &str = "tls";
+    static EXAMPLE_CA_CERTS: &str = "certs/ca.pem";
+    static EXAMPLE_CLIENT_CERT: &str = "certs/client.crt";
+    static EXAMPLE_CLIENT_KEY: &str = "certs/client.key";
+    static EXAMPLE_SERVER_CERT: &str = "certs/server.crt";
+    static EXAMPLE_SERVER_KEY: &str = "certs/server.key";
+    static EXAMPLE_SERVICE_ENDPOINT: &str = "127.0.0.1:8043";
+    static EXAMPLE_NETWORK_ENDPOINT: &str = "127.0.0.1:8044";
+    static EXAMPLE_NODE_ID: &str = "012";
+
+    /// Asserts config values based on the TEST_TOML file values.
+    fn assert_config_values(config: PartialConfig) {
+        assert_eq!(config.storage(), Some(EXAMPLE_STORAGE.to_string()));
+        assert_eq!(config.transport(), Some(EXAMPLE_TRANSPORT.to_string()));
+        assert_eq!(config.cert_dir(), None);
+        assert_eq!(config.ca_certs(), Some(EXAMPLE_CA_CERTS.to_string()));
+        assert_eq!(config.client_cert(), Some(EXAMPLE_CLIENT_CERT.to_string()));
+        assert_eq!(config.client_key(), Some(EXAMPLE_CLIENT_KEY.to_string()));
+        assert_eq!(config.server_cert(), Some(EXAMPLE_SERVER_CERT.to_string()));
+        assert_eq!(config.server_key(), Some(EXAMPLE_SERVER_KEY.to_string()));
+        assert_eq!(
+            config.service_endpoint(),
+            Some(EXAMPLE_SERVICE_ENDPOINT.to_string())
+        );
+        assert_eq!(
+            config.network_endpoint(),
+            Some(EXAMPLE_NETWORK_ENDPOINT.to_string())
+        );
+        assert_eq!(config.peers(), Some(vec![]));
+        assert_eq!(config.node_id(), Some(EXAMPLE_NODE_ID.to_string()));
+        assert_eq!(config.bind(), None);
+        #[cfg(feature = "database")]
+        assert_eq!(config.database(), None);
+        assert_eq!(config.registry_backend(), None);
+        assert_eq!(config.registry_file(), None);
+        assert_eq!(config.heartbeat_interval(), None);
+        assert_eq!(config.admin_service_coordinator_timeout(), None);
+    }
+
+    #[cfg(not(feature = "config-toml"))]
+    #[test]
+    /// This test verifies that a PartialConfig object, constructed from the TEST_TOML file using
+    /// PartialConfig module's `from_file` method, contains the correct values using the following
+    /// steps:
+    ///
+    /// 1. The example config toml file, TEST_TOML, is opened.
+    /// 2. A PartialConfig object is created by passing the opened file into the `from_file`
+    ///    function defined in the PartialConfig module.
+    ///
+    /// This test then verifies the PartialConfig object built in step 2 contains the correct
+    /// values by asserting each expected value.
+    fn test_partial_config_from_file() {
+        // Opening the toml file using the TEST_TOML path
+        let config_file =
+            fs::File::open(TEST_TOML).expect(&format!("Unable to load {}", TEST_TOML));
+        // Use the TomlConfig module's `from_file` method to construct a PartialConfig object
+        // from the config file previously opened.
+        let generated_config = from_file(config_file).unwrap();
+        // Compare the generated PartialConfig object against the expected values.
+        assert_config_values(generated_config);
+    }
+
+    #[cfg(feature = "config-toml")]
+    #[test]
+    /// This test verifies that a PartialConfig object, constructed from the TomlConfig module,
+    /// contains the correct values using the following steps:
+    ///
+    /// 1. The example config toml file, TEST_TOML, is read and converted to a string.
+    /// 2. A TomlConfig object is constructed by passing in the toml string created in the previous
+    ///    step.
+    /// 3. The TomlConfig object is transformed to a PartialConfig object using the `build` method.
+    ///
+    /// This test then verifies the PartialConfig object built from the TomlConfig object by
+    /// asserting each expected value.
+    fn test_toml_build() {
+        // Read the TEST_TOML example file to a string.
+        let toml_string =
+            fs::read_to_string(TEST_TOML).expect(&format!("Unable to load {}", TEST_TOML));
+        // Create a TomlConfig object from the toml string.
+        let toml_builder = TomlConfig::new(toml_string)
+            .expect(&format!("Unable to create TomlConfig from: {}", TEST_TOML));
+        // Build a PartialConfig from the TomlConfig object created.
+        let built_config = toml_builder.build();
+        // Compare the generated PartialConfig object against the expected values.
+        assert_config_values(built_config);
+    }
+}
